requirements.txt
# Since the requirements.txt file doesn't support the --trusted-host option
# (https://pip.pypa.io/en/stable/reference/pip_install/#requirements-file-format),
# you may need to use the --extra-index-url option instead.

appdirs==1.4.3
arrow==0.17.0
certifi
cffi>=1.9.1
click==6.7
configparser==4.0.2
coverage==4.5.2
cryptography==3.2.1
httpsig-cffi==15.0.0
Jinja2==2.10.1
jmespath==0.10.0
ndg-httpsclient==0.4.2
mock==2.0.0
oci==2.27.0
packaging==20.2
pluggy==0.4.0
py==1.4.33
pyasn1==0.2.3
pyOpenSSL==19.1.0
pycparser==2.20
pyparsing==2.2.0
pytest==3.2.3
pytest-cov==2.5.1
python-dateutil>=2.5.3,<3.0.0
pytz>=2016.10
requests==2.21.0
retrying==1.3.3
setuptools<45
six==1.14.0
sphinx==1.6.4
sphinx-rtd-theme==0.2.5b1
terminaltables==3.1.0
tox==2.9.1
vcrpy==1.13.0
virtualenv==15.1.0
pytest-xdist==1.22.2
pytest-forked==1.0.2
PyYAML==5.1.2
